requirements.txt
core==1.0.1
dotdict==0.1
dv==1.0.10
h5py==3.6.0
kornia==0.6.5
matplotlib==3.5.1
numpy==1.21.5
opencv_python==4.6.0.66
pandas==1.4.2
pathlib2==2.3.7.post1
Pillow==10.0.0
plotly==5.6.0
pytorch_lightning==1.6.4
PyYAML==6.0.1
scikit_learn==1.3.0
scipy==1.7.3
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
